+++ webots_ros2_core/CHANGELOG.rst
@@ -2,6 +2,10 @@
 Changelog for package webots_ros2_core
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+1.0.0 (2020-09-01)
+------------------
+* Added a universal 'webots_differential_drive_node' node
+
 0.0.4 (2020-07-03)
 ------------------
 * Fixed dependencies issue.
@@ -9,10 +13,10 @@ Changelog for package webots_ros2_core
 0.0.3 (2020-06-15)
 ------------------
 * Added support for multi robots.
-* Added a new tf_publisher node to publish tf of desired Webots nodes.
+* Added a new TfPublisher class to publish transforms of all the Solid nodes of the robot (if the robot `supervisor` field is true).
 * Added the possibility to run nodes in synchronized mode (using the 'synchronization' parameter).
-* Added a laser_publisher module.
 * Added better support for differential drive robots ('WebotsDifferentialDriveNode' class).
+* Added CameraDevice, LEDDevice and LaserDevice that create correspoding ROS2 topics
 
 0.0.2 (2019-09-23)
 ------------------