+++ webots_ros2_examples/CHANGELOG.rst
@@ -2,6 +2,10 @@
 Changelog for package webots_ros2_examples
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+1.0.0 (2020-09-01)
+------------------
+* Added a khepera_driver node
+
 0.0.3 (2020-06-15)
 ------------------
 * Added a new subscriber method to the `cmd_vel` topic to control the robot velocity.