+++ webots_ros2_universal_robot/CHANGELOG.rst
@@ -2,6 +2,10 @@
 Changelog for package webots_ros2_universal_robot
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+1.0.0 (2020-09-01)
+------------------
+* Use the webots_ros2_core::webots_robotic_arm_node node
+
 0.0.3 (2020-06-15)
 ------------------
 * Added an 'universal_robot_multiple' simulation and launch file.